leave precip and hour_multiplier out of pyeto

+++ custom_components/smart_irrigation/calcmodules/calcmodule.py
@@ -46,7 +46,7 @@ class SmartIrrigationCalculationModule:
         """Return the description of the calculation module."""
         return self._description
 
-    def calculate(self) -> int:
+    def calculate(self) -> float:
         """Perform the calculation and return the result as an integer."""
 
     def schema_serialized(self):
